% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dctree.R
\name{compute_mst}
\alias{compute_mst}
\title{Compute Minimum Spanning Tree using Prim's algorithm}
\usage{
compute_mst(dist_matrix)
}
\arguments{
\item{dist_matrix}{A symmetric distance matrix}
}
\value{
An igraph MST object
}
\description{
Compute Minimum Spanning Tree using Prim's algorithm
}
