% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dctree.R
\name{dc_distances}
\alias{dc_distances}
\title{Compute DC distances for a DCTree object}
\usage{
dc_distances(dctree)
}
\arguments{
\item{dctree}{A DCTree object}
}
\value{
A matrix of DC-distances
}
\description{
Generic method to compute DC-distances from a DCTree object.
}
