% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disco.R
\docType{package}
\name{disco-package}
\alias{disco}
\alias{disco-package}
\title{disco: Density-Based Internal Score for Clustering Outcomes}
\description{
Implementation of DISCO (Density-based Internal Score for Clustering Outcomes), a cluster validity index for evaluating density-based clustering results. DISCO is the first CVI that explicitly evaluates the quality of noise labels in addition to cluster quality. It assesses compactness and separation using density-connectivity distance while providing interpretable, bounded scores.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/aminentezari/Disco-R}
  \item Report bugs at \url{https://github.com/aminentezari/Disco-R/issues}
}

}
\author{
\strong{Maintainer}: Amin Entezari \email{amin_entezari@outlook.com}

Authors:
\itemize{
  \item Anna Beer (Original Python implementation)
  \item Lena Krieger (Original Python implementation)
  \item Pascal Weber (Original Python implementation)
}

}
\keyword{internal}
