% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot_disco_scores}
\alias{plot_disco_scores}
\title{Plot DISCO scores}
\usage{
plot_disco_scores(X, labels, scores = NULL, main = "DISCO Scores", ...)
}
\arguments{
\item{X}{A numeric matrix with 2 columns for plotting}

\item{labels}{Integer vector of cluster labels}

\item{scores}{Optional numeric vector of DISCO scores to color points}

\item{main}{Character, plot title. Default is "DISCO Scores".}

\item{...}{Additional arguments passed to plot()}
}
\description{
Visualize pointwise DISCO scores with color coding.
}
\examples{
\dontrun{
data <- make_moons(n_samples = 200)
scores <- disco_samples(data$X, data$labels)
plot_disco_scores(data$X, data$labels, scores)
}
}
