% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dctree.R
\name{print.DCTree}
\alias{print.DCTree}
\title{Print method for DCTree objects}
\usage{
\method{print}{DCTree}(x, ...)
}
\arguments{
\item{x}{A DCTree object}

\item{...}{Additional arguments (ignored)}
}
\description{
Print method for DCTree objects
}
