% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{validate_disco_input}
\alias{validate_disco_input}
\title{Validate DISCO input}
\usage{
validate_disco_input(X, labels, min_points)
}
\arguments{
\item{X}{Data matrix}

\item{labels}{Label vector}

\item{min_points}{Minimum points parameter}
}
\value{
TRUE if valid, otherwise stops with error
}
\description{
Check if inputs to DISCO functions are valid.
}
\keyword{internal}
