% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disco.R
\docType{package}
\name{disco-package}
\alias{disco-package}
\alias{disco}
\title{disco: Density-Informed Clustering Scoring and Optimisation}
\description{
The DISCO metric evaluates the quality of a clustering result using
density-connectivity (DC) distances derived from the minimum spanning tree
of the mutual-reachability graph.  It generalises the silhouette coefficient
to density-based partitions and assigns a principled score to noise points.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/aminentezari/Disco-R}
  \item Report bugs at \url{https://github.com/aminentezari/Disco-R/issues}
}

}
\author{
\strong{Maintainer}: Amin Entezari \email{amin_entezari@outlook.com}

Authors:
\itemize{
  \item Anna Beer (Original Python implementation)
  \item Lena Krieger (Original Python implementation)
  \item Pascal Weber (Original Python implementation)
}

}
