% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dctree.R
\name{extract_dc_distances_from_mst}
\alias{extract_dc_distances_from_mst}
\title{Extract DC-distances from MST}
\usage{
extract_dc_distances_from_mst(mst_graph, n)
}
\arguments{
\item{mst_graph}{An igraph MST object}

\item{n}{Number of nodes}
}
\value{
A matrix of DC-distances
}
\description{
DC-distance is the minimax path distance in the MST (maximum edge weight 
along the unique path between two nodes).
}
