% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disco.R
\name{p_cluster}
\alias{p_cluster}
\title{Compute p_cluster score for cluster points}
\usage{
p_cluster(X, labels, min_points = 5, precomputed_dc_dists = FALSE)
}
\arguments{
\item{X}{Either a numeric matrix/data.frame of features, or a distance matrix 
if precomputed_dc_dists=TRUE}

\item{labels}{Integer vector of cluster labels}

\item{min_points}{Integer, minimum number of points for dc-distance. Default is 5.}

\item{precomputed_dc_dists}{Logical, whether X is already a dc-distance matrix. 
Default is FALSE.}
}
\value{
A numeric vector of p_cluster scores for each sample
}
\description{
p_cluster is similar to the Silhouette Coefficient but uses dc-distance metric.
It evaluates cluster compactness and separation.

Note: -1 is NOT handled as noise here, but as a valid cluster label!
}
